게시물 삭제시 권한 체크

+++ src/main/java/com/sbs/webp/lolHi/controller/usr/ArticleController.java
@@ -179,6 +179,15 @@ public class ArticleController {
 			return "common/redirect";
 		}
 		
+		Article article = articleService.getArticleById(id);
+		
+		if (article.getMemberId() != loginedMemberId) {
+			model.addAttribute("msg", "삭제 권한이 없습니다.");
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirect";
+		}
+		
 		articleService.deleteArticleById(id);
 		
 		model.addAttribute("msg", String.format("%d번 게시물이 삭제되었습니다.", id));
